requirements.txt
pytorch
torchvision
cython
numpy
matplotlib
pandas
scipy
opencv
h5py
tqdm
cupy
ipdb